requirements.txt
Flask==0.10.1
gunicorn==19.3.0
elasticsearch==1.4.0
python-geoip==1.2
python-geoip-geolite2==2015.0303
Flask-WTF==0.11
Flask-Login==0.2.11
pusher==0.8
Flask-PyMongo==0.3.1